0.0.3
- Added BorealisOS Service
- Added BorealisOS Serverside Plugins
- Added authentication flow for Spotify Plugin
- Added automatic CEF debugging activation
- Improved Plugin Logging
- Continued working on update script

+++ install.sh
@@ -10,7 +10,16 @@ echo "                                                   ";
 echo "A SteamOS 3.0 plugin and customisation framework built using NodeJS. ";
 echo
 echo
-echo
+
+printf 'Are you sure you want to install (y/n)? '
+old_stty_cfg=$(stty -g)
+stty raw -echo
+answer=$( while ! head -c 1 | grep -i '[ny]' ;do true ;done )
+stty $old_stty_cfg
+if echo "$answer" | grep -iq "^n" ;then
+    exit 0
+fi
+
 
 # Clean Previous installation
 rm -r ~/.local/share/borealisOS/
@@ -18,6 +27,10 @@ rm -r ~/.local/share/borealisOS/
 sed -i '/# Added by BorealisOS/d' ~/.bashrc
 
 mkdir -p ~/.local/share/borealisOS
+
+systemctl stop borealisOS 2> /dev/null
+systemctl disable borealisOS 2> /dev/null
+rm -f /etc/systemd/system/borealisOS.service
 
 # Clone BorealisOS
 curl -L https://github.com/PineappleIOnic/BorealisOS/tarball/main | tar -xz -C ~/.local/share/borealisOS/ --strip-components=1
@@ -38,10 +51,13 @@ cd ~/.local/share/borealisOS
 mkdir -p ~/bin
 touch ~/bin/borealisOS
 
+# Install Service Files
+
+
 echo "cd ~/.local/share/borealisOS/ && ~/.local/share/borealisOS/node/bin/node src/index.js" > ~/bin/borealisOS
 chmod +x ~/bin/borealisOS
 
 # Add to $PATH
 echo "export PATH=\$PATH:~/bin # Added by BorealisOS" >> ~/.bashrc
 
-echo "BorealisOS was successfully installed! You can now run BorealisOS by typing 'borealisOS' in your terminal while Gaming mode is active."
+echo "BorealisOS was successfully installed! Reload into gaming mode to see the changes."
